fix(i18n): split locale catalogs and address review feedback
Move EN/zh-CN strings into dedicated message modules, derive the runtime
message map from the English catalog, restore browse CTA styling and
singular word counts, and wire emitters to catalog strings to prevent drift.
