+++ electron/locale/en.ts
@@ -1,0 +1,43 @@
+export const en = {
+  file: "File",
+  openProject: "Open Project…",
+  reopenLast: "Reopen Last Project",
+  recentProjects: "Recent Projects",
+  noRecent: "No Recent Projects",
+  clearRecent: "Clear Recent Projects",
+  edit: "Edit",
+  undo: "Undo",
+  redo: "Redo",
+  cut: "Cut",
+  copy: "Copy",
+  paste: "Paste",
+  pasteMatch: "Paste and Match Style",
+  delete: "Delete",
+  selectAll: "Select All",
+  view: "View",
+  reload: "Reload",
+  forceReload: "Force Reload",
+  devTools: "Toggle Developer Tools",
+  resetZoom: "Actual Size",
+  zoomIn: "Zoom In",
+  zoomOut: "Zoom Out",
+  fullscreen: "Toggle Full Screen",
+  window: "Window",
+  minimize: "Minimize",
+  zoom: "Zoom",
+  front: "Bring All to Front",
+  close: "Close Window",
+  quit: "Quit Rescript",
+  about: "About Rescript",
+  services: "Services",
+  hide: "Hide Rescript",
+  hideOthers: "Hide Others",
+  unhide: "Show All",
+  restart: "Restart",
+  later: "Later",
+  updateTitle: "Update available",
+  updateMessage: "Rescript {version} is ready to install.",
+  updateDetail: "Restart now to apply the update.",
+} as const;
+
+export type DesktopMessageKey = keyof typeof en;
