+++ electron/locale/zh-CN.ts
@@ -1,0 +1,43 @@
+import type { DesktopMessageKey } from "./en";
+
+export const zhCN: Record<DesktopMessageKey, string> = {
+  file: "文件",
+  openProject: "打开项目…",
+  reopenLast: "重新打开上一个项目",
+  recentProjects: "最近项目",
+  noRecent: "没有最近项目",
+  clearRecent: "清除最近项目",
+  edit: "编辑",
+  undo: "撤销",
+  redo: "重做",
+  cut: "剪切",
+  copy: "复制",
+  paste: "粘贴",
+  pasteMatch: "粘贴并匹配样式",
+  delete: "删除",
+  selectAll: "全选",
+  view: "视图",
+  reload: "重新加载",
+  forceReload: "强制重新加载",
+  devTools: "切换开发者工具",
+  resetZoom: "实际大小",
+  zoomIn: "放大",
+  zoomOut: "缩小",
+  fullscreen: "切换全屏",
+  window: "窗口",
+  minimize: "最小化",
+  zoom: "缩放",
+  front: "前置全部窗口",
+  close: "关闭窗口",
+  quit: "退出 Rescript",
+  about: "关于 Rescript",
+  services: "服务",
+  hide: "隐藏 Rescript",
+  hideOthers: "隐藏其他应用",
+  unhide: "全部显示",
+  restart: "立即重启",
+  later: "稍后",
+  updateTitle: "有可用更新",
+  updateMessage: "Rescript {version} 已准备好安装。",
+  updateDetail: "立即重启以应用更新。",
+};
