--- lib/i18n.ts
@@ -1,542 +1,0 @@
-export type UiLocale = "en" | "zh-CN";
-export type UiLocalePreference = "system" | UiLocale;
-
-export const DEFAULT_UI_LOCALE_PREFERENCE: UiLocalePreference = "system";
-export const UI_LOCALE_STORAGE_KEY = "rescript.ui-locale";
-
-const en = {
-  "app.title": "Rescript — edit videos like you edit text",
-  "common.cancel": "Cancel",
-  "common.close": "Close",
-  "common.delete": "Delete",
-  "common.download": "Download",
-  "common.import": "Import",
-  "common.loading": "Loading",
-  "common.remove": "Remove",
-  "common.restore": "Restore",
-  "common.retry": "Try again",
-  "common.searchOrCreate": "Search or create…",
-  "common.settings": "Settings",
-  "common.system": "System",
-  "common.tools": "Tools",
-  "settings.appearance": "Appearance",
-  "settings.light": "Light",
-  "settings.dark": "Dark",
-  "settings.interfaceLanguage": "Interface language",
-  "settings.privacy": "Privacy",
-  "settings.helpImprove": "Help improve the app",
-  "settings.telemetryHelp": "Send anonymous feature usage statistics and crash reports.",
-  "settings.support": "Support / feedback",
-  "settings.reportIssue": "Report an issue",
-  "settings.homepage": "Homepage",
-  "settings.github": "GitHub",
-  "settings.followX": "Follow on X",
-  "language.english": "English",
-  "language.simplifiedChinese": "简体中文",
-  "model.transcriptSource": "Transcript source",
-  "model.language": "Language",
-  "model.transcriptLanguage": "Transcript language",
-  "model.importTranscript": "Import transcript",
-  "model.baseDescription": "Faster download and transcription. Good for most clips.",
-  "model.smallDescription": "More accurate on longer or noisier audio. Larger download.",
-  "model.parakeetDescription": "NVIDIA FastConformer — faster on WebGPU, strong EU-language accuracy. Auto-detects language.",
-  "upload.recentProjects": "Recent projects",
-  "upload.removeRecent": "Remove from recent",
-  "upload.dropOrBrowse": "Drop a video or audio file, or {browse}",
-  "upload.browse": "browse",
-  "upload.chooseTranscriptFirst": "Pick a transcript in the menu above, then drop your media",
-  "upload.willUseTranscript": "Will use {name} · MP4, WebM, MOV, MP3, WAV, …",
-  "upload.mediaFormats": "MP4, WebM, MOV, MP3, WAV, M4A, …",
-  "upload.gettingReady": "Getting things ready",
-  "upload.gettingReadyHelp": "Preparing the local media engine…",
-  "upload.unsupported": "This browser cannot run the local media engine",
-  "upload.unsupportedHelp": "Use a current Chromium-based browser or the desktop app.",
-  "upload.transcribeTitle": "Transcribe",
-  "upload.transcribeText": "Whisper locally, or import SRT / VTT.",
-  "upload.editTitle": "Edit",
-  "upload.editText": "Select words and hit delete to edit.",
-  "upload.exportTitle": "Export",
-  "upload.exportText": "Render the final cut to MP4 or M4A.",
-  "upload.justNow": "just now",
-  "editor.chooseTranscript": "Choose a transcript file from the source menu first.",
-  "editor.chooseMedia": "Please choose a video or audio file.",
-  "editor.undo": "Undo (⌘Z)",
-  "editor.redo": "Redo (⇧⌘Z)",
-  "editor.export": "Export",
-  "topbar.startOver": "Start over",
-  "transcript.replace": "Replace transcript from SRT, VTT, or JSON",
-  "transcript.header": "Transcript",
-  "transcript.wordsDeleted": "{count} words",
-  "transcript.replaceConfirm": "Replace the current transcript with this file?",
-  "transcript.invalidFile": "Choose an SRT, VTT, or JSON transcript file.",
-  "transcript.noSpeech": "No speech detected. Make sure this file has audio, or import a transcript.",
-  "transcript.cut": "Cut",
-  "transcript.follow": "Follow playhead",
-  "transcript.hideDeleted": "Hide deleted words",
-  "transcript.showDeleted": "Show deleted words",
-  "transcript.correct": "Correct",
-  "transcript.scrollWithPlayhead": "Scroll with the playhead",
-  "transcript.joinSplit": "Clip split — click to join these clips",
-  "transcript.joinClips": "Join clips",
-  "transcript.hesitation": "Detected hesitation (not transcribed) — cut with Remove filler words",
-  "tools.bulk": "Bulk transcript cleanups",
-  "tools.removeFillers": "Remove filler words",
-  "tools.removeFillersTitle": "Cut filler words (\"um\", \"uh\", \"...\", …) from the media",
-  "tools.restoreFillers": "Restore filler words",
-  "tools.restoreFillersTitle": "Bring every cut filler word back",
-  "tools.removeSilences": "Remove silences",
-  "tools.removeSilencesTitle": "Cut pauses and silences (≥{seconds}s) from the media",
-  "tools.restoreSilences": "Restore silences",
-  "tools.restoreSilencesTitle": "Bring every cut pause back",
-  "timeline.removed": "{trimmed} removed — original length {duration}",
-  "timeline.back5": "Back 5 s",
-  "timeline.playPause": "Play / pause (space)",
-  "timeline.forward5": "Forward 5 s",
-  "timeline.split": "Split",
-  "timeline.splitTitle": "Split clip at playhead (S)",
-  "timeline.splitDisabled": "Move the playhead onto a kept region to split",
-  "timeline.delete": "Delete",
-  "timeline.deleteTitle": "Delete selected clip (Delete)",
-  "timeline.deleteDisabled": "Select a clip on the timeline to delete",
-  "timeline.restore": "Restore",
-  "timeline.restoreTitle": "Restore selected cut (Delete)",
-  "timeline.restoreDisabled": "Select a cut or silence section on the timeline to restore",
-  "timeline.zoomOut": "Zoom out",
-  "timeline.fit": "Fit to window",
-  "timeline.zoomIn": "Zoom in — drag word edges to refine timing",
-  "timeline.joinClips": "Join these clips (remove split)",
-  "timeline.trimStart": "Trim clip start",
-  "timeline.trimEnd": "Trim clip end",
-  "timeline.hesitationAdjust": "… detected hesitation — drag edges to adjust, or Remove filler words to cut",
-  "timeline.hesitationCut": "… detected hesitation — cut with Remove filler words",
-  "timeline.dragTiming": "{word} — drag edges to adjust timing",
-  "timeline.scrollZoom": "Scroll to zoom in/out",
-  "speaker.moveStart": "Drag to move where this speaker starts",
-  "speaker.moveLabel": "Move speaker label",
-  "speaker.options": "Speaker options",
-  "speaker.change": "Change speaker",
-  "speaker.rename": "Rename speaker",
-  "speaker.renameAction": "Rename",
-  "speaker.replace": "Replace in project with…",
-  "speaker.removeProject": "Remove from project",
-  "speaker.newName": "New name",
-  "speaker.find": "Find a speaker…",
-  "speaker.noMatches": "No matching speakers",
-  "speaker.renameTo": "Rename {current} to {next}",
-  "speaker.button": "Speaker",
-  "speaker.search": "Search or create…",
-  "speaker.defaultName": "Speaker {number}",
-  "speaker.create": "Create \"{name}\"",
-  "export.title": "Export",
-  "export.type": "Export type",
-  "export.video": "Video",
-  "export.audio": "Audio",
-  "export.transcript": "Transcript",
-  "export.subtitles": "Subtitles",
-  "export.videoUnavailable": "Video export isn’t available for audio-only projects",
-  "export.noAudio": "This file has no audio track",
-  "export.noWordsFirst": "Transcribe or import a transcript first",
-  "export.format": "Format",
-  "export.resolution": "Resolution",
-  "export.original": "Original",
-  "export.plainText": "Plain text",
-  "export.statOriginal": "Original",
-  "export.statCuts": "Cuts",
-  "export.statEdited": "Edited",
-  "export.transcriptHelp": "Speaker-labeled text with cuts removed. No timestamps.",
-  "export.subtitlesHelp": "SRT and VTT use the edited timeline with cuts applied. JSON keeps the full word list for re-import.",
-  "export.encodingHelp": "Re-encoding on your device — longer files take a while.",
-  "export.reexport": "Re-export with latest edits",
-  "export.rendering": "Rendering on your device…",
-  "export.downloadFile": "Download {name}",
-  "export.downloadFormat": "Download .{format}",
-  "export.exportFormat": "Export {format}",
-  "export.tryAgain": "Try again",
-  "import.reading": "Reading transcript…",
-  "import.importing": "Import…",
-  "import.failed": "Import failed",
-  "import.readingFile": "Reading file…",
-  "import.chooseFileShort": "Choose a file…",
-  "import.chooseFile": "Choose an SRT, VTT, or JSON transcript",
-  "import.selected": "{name} selected",
-  "banner.faster": "Want faster transcription and exports?",
-  "banner.getDesktop": "Get the Rescript desktop app",
-  "banner.downloadFor": "Download for {platform}",
-  "banner.dismiss": "Dismiss",
-  "social.githubRepo": "GitHub repository",
-  "social.discordServer": "Discord server",
-  "social.xProfile": "X profile",
-  "globalError.title": "Rescript — something went wrong",
-  "globalError.heading": "Something went wrong",
-  "globalError.body": "The editor hit an unexpected error. Your saved projects are still on this device — reloading should bring them back.",
-  "progress.loadingMedia": "Loading media…",
-  "progress.loadingMediaEngine": "Loading media engine…",
-  "progress.extractingAudio": "Extracting audio…",
-  "progress.loadingSpeechModel": "Loading speech model…",
-  "progress.loadingSpeechCache": "Loading speech model from cache…",
-  "progress.downloadingSpeech": "Downloading speech model…",
-  "progress.gpuFallback": "GPU interrupted — continuing on CPU…",
-  "progress.detectingSpeech": "Detecting speech…",
-  "progress.detectingLanguage": "Detecting source language…",
-  "progress.transcribing": "Transcribing…",
-  "progress.loadingAlignCache": "Loading alignment model from cache…",
-  "progress.downloadingAlign": "Downloading alignment model…",
-  "progress.aligning": "Aligning words…",
-  "progress.speakers": "Identifying speakers…",
-  "error.selectModel": "Select a speech model to transcribe.",
-  "error.workerCrashed": "Transcription worker crashed.",
-  "error.mediaEngineNetwork": "Couldn't load the media engine — the connection dropped. Check your internet and try again.",
-  "error.processFile": "Failed to process this file.",
-  "error.extractAudio": "Could not extract audio from this file.",
-  "error.nothingToExport": "Everything has been deleted — nothing to export.",
-  "error.videoExport": "Export failed while rendering the video.",
-  "error.audioExport": "Export failed while rendering the audio.",
-  "error.export": "Export failed.",
-  "error.emptyTranscript": "That transcript file is empty.",
-  "error.noTimedWords": "No timed words found in that transcript.",
-  "error.parseJson": "Could not parse that JSON transcript.",
-  "error.jsonShape": "JSON must be a word array or an object containing words.",
-  "error.noWords": "No words to export.",
-  "error.projectMissing": "That project is no longer saved.",
-  "error.openProject": "Could not open that project.",
-  "error.removeProject": "Could not remove that project.",
-  "error.readTranscript": "Could not read that transcript.",
-  "error.clearRecent": "Could not clear recent projects.",
-  "error.modelDownload": "Couldn't finish downloading the speech model — the connection dropped. Check your internet and try again; completed parts are kept.",
-  "error.gpuReset": "Transcription was interrupted when the GPU reset, often after locking the screen. Please try again.",
-  "confirm.clearRecent": "Remove all recent projects? Their saved edits are deleted.",
-} as const;
-
-export type MessageKey = keyof typeof en;
-
-const zhCN: Record<MessageKey, string> = {
-  "app.title": "Rescript — 像编辑文字一样编辑视频",
-  "common.cancel": "取消",
-  "common.close": "关闭",
-  "common.delete": "删除",
-  "common.download": "下载",
-  "common.import": "导入",
-  "common.loading": "正在加载",
-  "common.remove": "移除",
-  "common.restore": "恢复",
-  "common.retry": "重试",
-  "common.searchOrCreate": "搜索或新建…",
-  "common.settings": "设置",
-  "common.system": "跟随系统",
-  "common.tools": "工具",
-  "settings.appearance": "外观",
-  "settings.light": "浅色",
-  "settings.dark": "深色",
-  "settings.interfaceLanguage": "界面语言",
-  "settings.privacy": "隐私",
-  "settings.helpImprove": "帮助改进应用",
-  "settings.telemetryHelp": "发送匿名功能使用统计和崩溃报告。",
-  "settings.support": "支持与反馈",
-  "settings.reportIssue": "报告问题",
-  "settings.homepage": "主页",
-  "settings.github": "GitHub",
-  "settings.followX": "在 X 上关注",
-  "language.english": "English",
-  "language.simplifiedChinese": "简体中文",
-  "model.transcriptSource": "转录来源",
-  "model.language": "语言",
-  "model.transcriptLanguage": "转录语言",
-  "model.importTranscript": "导入转录文本",
-  "model.baseDescription": "下载和转录更快，适合大多数片段。",
-  "model.smallDescription": "较长或噪声较多的音频更准确，但下载体积更大。",
-  "model.parakeetDescription": "NVIDIA FastConformer；WebGPU 下更快，擅长欧洲语言并自动识别语种。",
-  "upload.recentProjects": "最近项目",
-  "upload.removeRecent": "从最近项目中移除",
-  "upload.dropOrBrowse": "拖入视频或音频文件，或者{browse}",
-  "upload.browse": "浏览文件",
-  "upload.chooseTranscriptFirst": "请先在上方菜单选择转录文本，再拖入媒体文件",
-  "upload.willUseTranscript": "将使用 {name} · MP4、WebM、MOV、MP3、WAV 等",
-  "upload.mediaFormats": "MP4、WebM、MOV、MP3、WAV、M4A 等",
-  "upload.gettingReady": "正在准备",
-  "upload.gettingReadyHelp": "正在准备本地媒体引擎…",
-  "upload.unsupported": "此浏览器无法运行本地媒体引擎",
-  "upload.unsupportedHelp": "请使用新版 Chromium 浏览器或桌面客户端。",
-  "upload.transcribeTitle": "转录",
-  "upload.transcribeText": "在本机运行 Whisper，或导入 SRT / VTT。",
-  "upload.editTitle": "编辑",
-  "upload.editText": "选择文字并按删除键即可剪除内容。",
-  "upload.exportTitle": "导出",
-  "upload.exportText": "将最终成片导出为 MP4 或 M4A。",
-  "upload.justNow": "刚刚",
-  "editor.chooseTranscript": "请先从转录来源菜单选择一个转录文本文件。",
-  "editor.chooseMedia": "请选择视频或音频文件。",
-  "editor.undo": "撤销 (⌘Z)",
-  "editor.redo": "重做 (⇧⌘Z)",
-  "editor.export": "导出",
-  "topbar.startOver": "重新开始",
-  "transcript.replace": "用 SRT、VTT 或 JSON 替换转录文本",
-  "transcript.header": "转录文本",
-  "transcript.wordsDeleted": "已删除 {count} 个词",
-  "transcript.replaceConfirm": "要用此文件替换当前转录文本吗？",
-  "transcript.invalidFile": "请选择 SRT、VTT 或 JSON 转录文本文件。",
-  "transcript.noSpeech": "未检测到语音。请确认文件包含音轨，或导入转录文本。",
-  "transcript.cut": "剪除",
-  "transcript.follow": "跟随播放位置",
-  "transcript.hideDeleted": "隐藏已删除文字",
-  "transcript.showDeleted": "显示已删除文字",
-  "transcript.correct": "纠正",
-  "transcript.scrollWithPlayhead": "跟随播放位置滚动",
-  "transcript.joinSplit": "片段分割点 — 点击以合并片段",
-  "transcript.joinClips": "合并片段",
-  "transcript.hesitation": "检测到未转录的停顿语气，可用“移除语气词”剪除",
-  "tools.bulk": "批量清理转录文本",
-  "tools.removeFillers": "移除语气词",
-  "tools.removeFillersTitle": "从媒体中剪除“嗯”“呃”“…”等语气词",
-  "tools.restoreFillers": "恢复语气词",
-  "tools.restoreFillersTitle": "恢复所有已剪除的语气词",
-  "tools.removeSilences": "移除静音",
-  "tools.removeSilencesTitle": "剪除不少于 {seconds} 秒的停顿和静音",
-  "tools.restoreSilences": "恢复静音",
-  "tools.restoreSilencesTitle": "恢复所有已剪除的停顿",
-  "timeline.removed": "已剪除 {trimmed} — 原始长度 {duration}",
-  "timeline.back5": "后退 5 秒",
-  "timeline.playPause": "播放 / 暂停（空格）",
-  "timeline.forward5": "前进 5 秒",
-  "timeline.split": "分割",
-  "timeline.splitTitle": "在播放位置分割片段（S）",
-  "timeline.splitDisabled": "请把播放位置移到保留区域后再分割",
-  "timeline.delete": "删除",
-  "timeline.deleteTitle": "删除所选片段（Delete）",
-  "timeline.deleteDisabled": "请先在时间轴上选择要删除的片段",
-  "timeline.restore": "恢复",
-  "timeline.restoreTitle": "恢复所选剪除区域（Delete）",
-  "timeline.restoreDisabled": "请先选择要恢复的剪除或静音区域",
-  "timeline.zoomOut": "缩小",
-  "timeline.fit": "适合窗口",
-  "timeline.zoomIn": "放大 — 拖动文字边缘可微调时间",
-  "timeline.joinClips": "合并片段（移除分割点）",
-  "timeline.trimStart": "修剪片段开头",
-  "timeline.trimEnd": "修剪片段结尾",
-  "timeline.hesitationAdjust": "检测到停顿语气 — 可拖动边缘调整，或用“移除语气词”剪除",
-  "timeline.hesitationCut": "检测到停顿语气 — 可用“移除语气词”剪除",
-  "timeline.dragTiming": "{word} — 拖动边缘可调整时间",
-  "timeline.scrollZoom": "滚动以缩放时间轴",
-  "speaker.moveStart": "拖动以调整此说话人的起始位置",
-  "speaker.moveLabel": "移动说话人标签",
-  "speaker.options": "说话人选项",
-  "speaker.change": "更改说话人",
-  "speaker.rename": "重命名说话人",
-  "speaker.renameAction": "重命名",
-  "speaker.replace": "在整个项目中替换为…",
-  "speaker.removeProject": "从项目中移除",
-  "speaker.newName": "新名称",
-  "speaker.find": "查找说话人…",
-  "speaker.noMatches": "没有匹配的说话人",
-  "speaker.renameTo": "将 {current} 重命名为 {next}",
-  "speaker.button": "说话人",
-  "speaker.search": "搜索或新建…",
-  "speaker.defaultName": "说话人 {number}",
-  "speaker.create": "新建“{name}”",
-  "export.title": "导出",
-  "export.type": "导出类型",
-  "export.video": "视频",
-  "export.audio": "音频",
-  "export.transcript": "转录文本",
-  "export.subtitles": "字幕",
-  "export.videoUnavailable": "纯音频项目不能导出视频",
-  "export.noAudio": "此文件没有音轨",
-  "export.noWordsFirst": "请先转录或导入转录文本",
-  "export.format": "格式",
-  "export.resolution": "分辨率",
-  "export.original": "原始",
-  "export.plainText": "纯文本",
-  "export.statOriginal": "原始",
-  "export.statCuts": "剪除",
-  "export.statEdited": "成片",
-  "export.transcriptHelp": "带说话人标签的文本，已移除剪除内容，不含时间戳。",
-  "export.subtitlesHelp": "SRT 和 VTT 使用已应用剪除的成片时间轴；JSON 保留完整词表以便重新导入。",
-  "export.encodingHelp": "正在本机重新编码，较长文件需要更多时间。",
-  "export.reexport": "按最新编辑重新导出",
-  "export.rendering": "正在本机渲染…",
-  "export.downloadFile": "下载 {name}",
-  "export.downloadFormat": "下载 .{format}",
-  "export.exportFormat": "导出 {format}",
-  "export.tryAgain": "重新导出",
-  "import.reading": "正在读取转录文本…",
-  "import.importing": "正在导入…",
-  "import.failed": "导入失败",
-  "import.readingFile": "正在读取文件…",
-  "import.chooseFileShort": "请选择文件…",
-  "import.chooseFile": "选择 SRT、VTT 或 JSON 转录文本",
-  "import.selected": "已选择 {name}",
-  "banner.faster": "想要更快地转录和导出？",
-  "banner.getDesktop": "获取 Rescript 桌面客户端",
-  "banner.downloadFor": "下载 {platform} 版",
-  "banner.dismiss": "关闭",
-  "social.githubRepo": "GitHub 仓库",
-  "social.discordServer": "Discord 服务器",
-  "social.xProfile": "X 个人主页",
-  "globalError.title": "Rescript — 出现错误",
-  "globalError.heading": "出现错误",
-  "globalError.body": "编辑器遇到了意外错误。已保存的项目仍在此设备上，重新加载后即可恢复。",
-  "progress.loadingMedia": "正在加载媒体…",
-  "progress.loadingMediaEngine": "正在加载媒体引擎…",
-  "progress.extractingAudio": "正在提取音频…",
-  "progress.loadingSpeechModel": "正在加载语音模型…",
-  "progress.loadingSpeechCache": "正在从缓存加载语音模型…",
-  "progress.downloadingSpeech": "正在下载语音模型…",
-  "progress.gpuFallback": "GPU 已中断，正在改用 CPU…",
-  "progress.detectingSpeech": "正在检测语音…",
-  "progress.detectingLanguage": "正在识别源语言…",
-  "progress.transcribing": "正在转录…",
-  "progress.loadingAlignCache": "正在从缓存加载对齐模型…",
-  "progress.downloadingAlign": "正在下载对齐模型…",
-  "progress.aligning": "正在对齐文字…",
-  "progress.speakers": "正在识别说话人…",
-  "error.selectModel": "请选择用于转录的语音模型。",
-  "error.workerCrashed": "转录进程发生崩溃。",
-  "error.mediaEngineNetwork": "媒体引擎加载失败，网络连接已中断。请检查网络后重试。",
-  "error.processFile": "无法处理此文件。",
-  "error.extractAudio": "无法从此文件提取音频。",
-  "error.nothingToExport": "所有内容都已删除，没有可导出的内容。",
-  "error.videoExport": "渲染视频时导出失败。",
-  "error.audioExport": "渲染音频时导出失败。",
-  "error.export": "导出失败。",
-  "error.emptyTranscript": "转录文本文件为空。",
-  "error.noTimedWords": "转录文本中没有带时间信息的文字。",
-  "error.parseJson": "无法解析 JSON 转录文本。",
-  "error.jsonShape": "JSON 必须是文字数组，或包含 words 字段的对象。",
-  "error.noWords": "没有可导出的文字。",
-  "error.projectMissing": "此项目已不在已保存项目中。",
-  "error.openProject": "无法打开此项目。",
-  "error.removeProject": "无法移除此项目。",
-  "error.readTranscript": "无法读取此转录文本。",
-  "error.clearRecent": "无法清除最近项目。",
-  "error.modelDownload": "语音模型下载未完成，网络连接已中断。请检查网络后重试；已下载完成的部分会被保留。",
-  "error.gpuReset": "GPU 重置导致转录中断（常见于锁屏后），请重试。",
-  "confirm.clearRecent": "要移除所有最近项目吗？已保存的编辑也会被删除。",
-};
-
-export type Translate = (
-  key: MessageKey,
-  params?: Record<string, string | number>
-) => string;
-
-export function isUiLocalePreference(value: unknown): value is UiLocalePreference {
-  return value === "system" || value === "en" || value === "zh-CN";
-}
-
-export function resolveUiLocale(
-  preference: UiLocalePreference,
-  systemLanguages: readonly string[]
-): UiLocale {
-  if (preference !== "system") return preference;
-  for (const raw of systemLanguages) {
-    const language = raw.toLowerCase();
-    if (language === "zh" || language.startsWith("zh-")) return "zh-CN";
-    if (language === "en" || language.startsWith("en-")) return "en";
-  }
-  return "en";
-}
-
-export function systemLanguages(): string[] {
-  if (typeof navigator === "undefined") return [];
-  return navigator.languages?.length
-    ? Array.from(navigator.languages)
-    : navigator.language
-      ? [navigator.language]
-      : [];
-}
-
-export function loadUiLocalePreference(): UiLocalePreference {
-  if (typeof window === "undefined") return DEFAULT_UI_LOCALE_PREFERENCE;
-  try {
-    const value = window.localStorage.getItem(UI_LOCALE_STORAGE_KEY);
-    if (isUiLocalePreference(value)) return value;
-  } catch {
-    // Private mode / disabled storage.
-  }
-  return DEFAULT_UI_LOCALE_PREFERENCE;
-}
-
-export function saveUiLocalePreference(preference: UiLocalePreference): void {
-  if (typeof window === "undefined") return;
-  try {
-    window.localStorage.setItem(UI_LOCALE_STORAGE_KEY, preference);
-  } catch {
-    // Private mode / disabled storage.
-  }
-}
-
-export function translate(
-  locale: UiLocale,
-  key: MessageKey,
-  params: Record<string, string | number> = {}
-): string {
-  const template = (locale === "zh-CN" ? zhCN : en)[key] ?? en[key];
-  return template.replace(/\{(\w+)\}/g, (token, name: string) =>
-    Object.prototype.hasOwnProperty.call(params, name) ? String(params[name]) : token
-  );
-}
-
-const runtimeMessageKeys: Record<string, MessageKey> = {
-  "Loading media…": "progress.loadingMedia",
-  "Loading media engine…": "progress.loadingMediaEngine",
-  "Extracting audio…": "progress.extractingAudio",
-  "Loading speech model…": "progress.loadingSpeechModel",
-  "Loading speech model from cache…": "progress.loadingSpeechCache",
-  "Downloading speech model…": "progress.downloadingSpeech",
-  "GPU interrupted — continuing on CPU…": "progress.gpuFallback",
-  "Detecting speech…": "progress.detectingSpeech",
-  "Detecting language…": "progress.detectingLanguage",
-  "Transcribing…": "progress.transcribing",
-  "Loading alignment model from cache…": "progress.loadingAlignCache",
-  "Downloading alignment model…": "progress.downloadingAlign",
-  "Aligning words…": "progress.aligning",
-  "Identifying speakers…": "progress.speakers",
-  "Select a speech model to transcribe.": "error.selectModel",
-  "Transcription worker crashed.": "error.workerCrashed",
-  "Couldn't load the media engine — the connection dropped. Check your internet and try again.": "error.mediaEngineNetwork",
-  "Failed to process this file.": "error.processFile",
-  "Could not extract audio from this file.": "error.extractAudio",
-  "Everything has been deleted — nothing to export.": "error.nothingToExport",
-  "Export failed while rendering the video.": "error.videoExport",
-  "Export failed while rendering the audio.": "error.audioExport",
-  "Export failed.": "error.export",
-  "That transcript file is empty.": "error.emptyTranscript",
-  "No timed words found in that transcript.": "error.noTimedWords",
-  "Could not parse that JSON transcript.": "error.parseJson",
-  'JSON must be a word array or { "words": [...] }.': "error.jsonShape",
-  "No words to export.": "error.noWords",
-  "That project is no longer saved.": "error.projectMissing",
-  "Could not open that project.": "error.openProject",
-  "Could not remove that project.": "error.removeProject",
-  "Could not read that transcript.": "error.readTranscript",
-  "Could not clear recent projects.": "error.clearRecent",
-  "Couldn't finish downloading the speech model — the connection dropped. Check your internet and try again; the parts that finished downloading are kept.": "error.modelDownload",
-  "Transcription was interrupted when the GPU reset (often after locking the screen). Please try again.": "error.gpuReset",
-};
-
-export function localizeRuntimeMessage(
-  text: string | null | undefined,
-  t: Translate
-): string {
-  if (!text) return "";
-  const key = runtimeMessageKeys[text];
-  return key ? t(key) : text;
-}
-
-export function formatRelativeTime(
-  locale: UiLocale,
-  timestamp: number,
-  now = Date.now()
-): string {
-  const seconds = Math.max(0, Math.round((now - timestamp) / 1000));
-  const formatter = new Intl.RelativeTimeFormat(locale, { numeric: "auto" });
-  if (seconds < 45) return formatter.format(0, "second");
-  const minutes = Math.round(seconds / 60);
-  if (minutes < 60) return formatter.format(-minutes, "minute");
-  const hours = Math.round(minutes / 60);
-  if (hours < 48) return formatter.format(-hours, "hour");
-  const days = Math.round(hours / 24);
-  if (days < 14) return formatter.format(-days, "day");
-  return new Intl.DateTimeFormat(locale, {
-    month: "short",
-    day: "numeric",
-  }).format(timestamp);
-}
