+++ lib/projects.ts
@@ -229,19 +229,3 @@ export function fileFromProject(project: ProjectRecord): File {
     lastModified: project.updatedAt,
   });
 }
-
-/** Compact relative time for the recent list (e.g. "just now", "3h ago"). */
-export function formatRelativeTime(ts: number, now = Date.now()): string {
-  const sec = Math.max(0, Math.round((now - ts) / 1000));
-  if (sec < 45) return "just now";
-  const min = Math.round(sec / 60);
-  if (min < 60) return `${min}m ago`;
-  const hr = Math.round(min / 60);
-  if (hr < 48) return `${hr}h ago`;
-  const day = Math.round(hr / 24);
-  if (day < 14) return `${day}d ago`;
-  return new Date(ts).toLocaleDateString(undefined, {
-    month: "short",
-    day: "numeric",
-  });
-}
